Added the VoltAcc class to accelerate the partial accross the gap to the wein filter Need to add the dimentions, enter and exit
also found issue with the origonal filter, need to re work the cord stuff in the instant function

+++ src/Model/Particle.java
@@ -13,11 +13,11 @@ public class Particle {
     protected Cord velocity;
     protected Cord acceleration;
 
-    float mass;
-    float charge;
+    double mass;
+    double charge;
 
 
-    public Particle( Cord position, Cord velocity,Cord acceleration, float mass, float charge) {
+    public Particle( Cord position, Cord velocity,Cord acceleration, double mass, double charge) {
         this.position = position;
         this.velocity = velocity;
         this.acceleration = acceleration;
